Net2.0: Rename package and path
