+++ p2pserver/message/message.go
@@ -27,7 +27,7 @@ import (
 	"fmt"
 
 	"github.com/ontio/ontology/common/log"
-	"github.com/ontio/ontology/net/protocol"
+	"github.com/ontio/ontology/p2pserver/protocol"
 )
 
 type Messager interface {
