+++ p2pserver/message/verack.go
@@ -24,7 +24,7 @@ import (
 	"strconv"
 
 	"github.com/ontio/ontology/common/log"
-	"github.com/ontio/ontology/net/protocol"
+	"github.com/ontio/ontology/p2pserver/protocol"
 )
 
 type verACK struct {
